pyproject.toml
[project]
name = "demo-mcp-learn"
version = "0.1.0"
description = "Multi-agent LOTR poetry demo using Microsoft Agent Framework and MCP"
requires-python = ">=3.10"
dependencies = [
    "agent-framework>=0.1.0b0",
    "mcp[cli]>=1.0.0b8",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.9",
    "pytest>=8.3.3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 100

[tool.pyright]
pythonVersion = "3.10"
